ginger-price-prediction/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
pandas
numpy
scikit-learn
requests
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0